Izzy - added and integrated arm
Added the arm script.
modified the player manager to accommodate the arm script.
imported the arm asset.
added new code for the hand collision.
added code for the wall and arm collisions.

NOTE:
Changed the curser from fixed update to update because the mouse axis update didn't seem to work well with it.